made User folder and moved Admin folder out of AUth
